requirements.txt
langchain
langchain-google-genai
PyMuPDF
python-dotenv
pandas
openpyxl
